chore: :recycle: Split local and vercel deploy entrypoint to not overwrite and push it to github
because then the vercel deploy always fail  if i forget to change it back
